Adding The CSS Code & Linking It in The HTML Code
The CSS code was added by creating a class selector (".selected") which will be added to the selected element.
This will make the elements background turn into green.

Furthermore, a "cursor:pointer;" rule was added to the <li> element to seem like the item lists are selectable.

Finally, the file "css/styles.css" was linked in the HTML file.