Implementing The Code For SHIFT + Click Events
This made deal with the imposter syndrome and code is a little messy!
It has the correct behaviour most of the time.

One case is needed to be corrected. If I select a single element and then
shift-click an element above it, it selects all the elements above it. If
I shift-click element below it afterwards, it keeps selected the initially selected
and all the element below it.

The problem raises when I first shift-select element below the initially selected
element, it selects all the element below it. No problem here.
And yet, if I shift-click an element above it, it keeps selected the previous elements
plus any element above it.